+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.0.75
 
 - [DSChip] Added a new height param to widget
+- [DSMessageItemModel] Added new customer map property
+- [DSCard] Added new customer map property; Improve Media view for tunnel users
 
 ## 0.0.74